refactor: Update route registration and clean up app structure
- Replaced the direct blueprint registration with a new `register_routes` function for better organization.
- Removed unused imports and commented-out code in `all_routes.py` to streamline the file.
- Consolidated blueprint registration for chat and inventory routes under a single function, enhancing maintainability.
